requirements.txt
configparser
pycrypto
